reload_page_while_failing method to reload pages that fail
The capybara wait mechanism is based on the concept that a page may take
a while to display the correct contents. In our scenario however we want
to continuously reload the page - as it will be waiting for content that
is being added to the content store.

This introduces a configuration value in RSpec which determines how many
times the page should be reloaded.