Refactor specialist publisher spec
"unsuccessfully creating a draft" was removed since it has been covered in unit testing and doesn't really represent a complete end to end test. Also refactored the before do block in the document attachment test and deleted the error method assertion in the assertion helper file since it is not required anymore.