Rake task to wait for router to be ready
When router starts up it initially returns 503 responses until it has a
single route in it's database. Once there is at least one route it will
return 404 for routes it does not have.

We want to wait for it to not respond to 503 so that our tests don't
need to be aware that a 503 is a valid waiting status.

We can also use this value as a means for determining whether to run
tests or not, as if we can't get a single item onto the router there
is probably a problem with the setup.