Add support for Signon
This commit adds support for logging in to Signon, to the existing
tests for Travel Advice Publisher and Specialist Publisher.

Without this, the way in which the applications under test were
configured (both with the Docker compose setup, and through
govuk-guix) was with the mock mode of the gds-sso gem. This means that
the first user in the database for each app is used, rather than using
Signon.

With this change, the existing tests can be run without this mock
mode. This has been tested with govuk-guix, were this functionality
also opens up the option of running the tests with the apps using the
Rails production environment.

The motivation for this is both to enable testing the apps in the
Rails production environment (which removes the option for using the
gds-sso mock mode), to remove some differences to how they are run in
the GOV.UK production environment, but also to enable writing new
tests that depend on testing with users with specific permissions, or
testing behaviour across apps that relies on Signon, for example,
creating and previewing access limited content.

To make this possible, the new SignonHelper module includes support
for navigating the sign in process, including two step verification if
necessary, and also the user admin interface for modifying user
permissions. Ideally as Signon isn't currently being tested, this
would be a bit easier, perhaps by directly using a HTTP API, but I
don't think Signon currently has a simple API to use.

Without setting the SIGNON_USER_COUNT environment variable, the
existing tests will work without Signon. However, if this is set, then
it should be an integer value that matches the number of users
specified in the SIGNON_USER_<number>_EMAIL and
SIGNON_USER_<number>_PASSPHRASE environment variables, where <number>
is replaced with the number of the user, starting from 0. One
additional requirement on these users, is that user 0 has the
superadmin role in Signon, as this user is used to configure the
permissions for the other users.