Wait for travel advice to be published before asserting
When publishing the initial travel advice for a country, the travel
advice publisher first sends over a version with a blank summary before
sending a version with the actual content. Related to 03e929d9

Occasionally the tests would pick up that blank summary document and
assert against that instead of the subsequent document.

Waiting for the content to be published remedies this race condition.