Add mainstream publisher to the list of apps booted up by docker-compose
Allows us to write end to end tests which involve the publishing of
content via mainstream publisher.

Mainstream publisher consists of the frontend, government-frontend and
publisher applications.

Publisher application depends on the calendar application which has thus
been pulled through.

The wait_for_router command accesses the / route which is now populated
by publish_special_routes.  Depend on the frontend and frontend-draft
from publishing-e2e-tests to ensure that the / route returns a 200.

By end to end testing mainstream publisher, this gives an additional layer
of confidence to proposed changes to these repos.