Add scenario for Publisher removing an artefact without a redirect
By adding this scenario for removing an artefact inside of mainstream
publisher we get confidence that publisher is correctly communicating
with the publishing-api to remove the content.

We also get confidence that the router has been set up to intercept the
gone resource.

We've added a test for the 404 behaviour of the subpage as documentation,
but appreciate that isn't correct.