Add scenario for Publisher updating an artefact in Frontend
By adding a scenario for updating a published artefact inside of mainstream
we get the confidence that the publisher is able to update an artefact.

Scheduling downtime for a transaction goes through a different flow
compared to the existing tests which go through the edit artefact
flow.  This gives us added confidence that the publisher downtime
flow is working.

When scheduling downtime for a transaction the message only appears
24 hours before the downtime is scheduled to begin.  By filling in
the start time to begin from now we are forcing the message to
display immediately.