Add scenario for Publisher removing an artefact and redirecting it
By adding this scenario we get confidence that publisher app
sends though the right redirection instructions to router.

The current behaviour of publisher is such that when removing an
artefact with subpages the redirect of the subpage will append the
subpage slug.  As far as we've been able to determine this is
correct and intended behaviour, however it is somewhat odd and
surprising.