Expect a 410, rather than 404 for discarded drafts
This change is a result of a change to the Content Store, that will
now delete routes in the router-api when the corresponding content item
is deleted.

The router defaults to converting deleting routes to gone routes,
which is something that I'm uneasy about, as it means that there is a
buildup of state in the router-api database, but that is an issue for
another time.

So, in summary, as the router-api chooses to respond with gones when
routes are deleted, and routes are now being deleted by the content
store when appropriate, the expectation should be changed from a
404 (not found) to a 410 (gone).