Add healthchecks for most of the applications
This seems to resolve a race condition on Jenkins where some of the
frontend applications are booted but no available as of the tests
running. Much like the other healthchecks these onse insure the
applications are not returning 500s on the given address before marking
the containers healthy meaning the tests shouldn't hit an unhealthy
container and fail.