Abort the build if not all repos were able to be cloned
It's a relatively common scenario that we can get failed tests due to
a commit we're trying to build for no longer existing - for example this
happens when someone force pushes. The aim of this is to make it easier
when looking through Jenkins failures to look for instances where tests
have actually failed or if we were just not able to build.

This adds a distinct try/catch around the cloning process and if that
fails the build is aborted.

Early attempts at solving this had the bin/clone script return a
specific exit code we could catch. However this proved problematic as we
run the clone process in parallel so it's hard to catch a single exit
code.