Retry opening the dialog modal until it appears
We've seen some flakeyness around the modal box not appearing after
the link to open it being clicked.  We believe this happens when the
link is clicked before bootstrap has registered the modal event handler.

Now we'll repeatedly open up the modal until the confirmation button
appears on the page.