Update Dockerfile and Makefile for running tests in production
We have an aim to run the test suite as close to production as
possible and this is the initial step of running and setting up
applications in the production environment with the caveat of no
Signon or SSL currently.

To allow for running the tests in production we added ENV vars and
ARGs to the underlying applications. These changes reflect those
variables and allow us to orchestrate the applications in
rails production env.