Proxy asset requests to fake S3 via Nginx
Currently the two feature flag percentages are set to 100% in
all environments [1]. This means that in those environments Asset
Manager proxies all asset requests to S3 via Nginx.

The tests in this repo do not take into account the environment
variables or the Nginx configuration defined in govuk-puppet. This meant
that the feature flag percentages were not set and so Asset Manager has
been using send_file to instruct Nginx to serve assets directly from the
filesystem. Furthermore, since the Nginx config was not sending the
relevant headers (and in any case was not setup to handle a send_file
response) the tests were relying on the Rails app serving the asset
directly from disk due to the absence of the X-Accel-Redirect request
header.

We want to make the proxying to S3 behaviour the default for Asset
Manager and plan to remove the ability to serve from the filesystem.

This commit sets the two feature flag percentages to 100% and adds just
enough Nginx configuration to have Asset Manager instruct Nginx to proxy
asset requests to a fake version of S3 provided by the Asset Manager
Rails app.

While this makes the tests more realistic and means we can safely remove
the old functionality from Asset Manager, I'm somewhat uncomfortable with
the fact that the Nginx config I've come up with differs somewhat from
that in all other environments.

Notably in those other environments, the assets-origin vhost proxies
requests to the asset-manager vhost and it's the latter which contains
the "internal" location used to proxy to S3; whereas in this repo the
assets-origin vhost proxies directly to the Asset Manager Rails app and
therefore itself contains the "internal" location. Similarly in the
other environments there is no need for the fake-s3 location at all,
because we can just rely on the default behaviour of the asset-manager
vhost.

I did attempt to make the Nginx config more realistic, but ran into
trouble because all the Nginx vhosts run in a single container which
doesn't seem to be able to resolve hosts like asset-manager.dev.gov.uk.

All in all the fact that this repo doesn't make use of the environment
variables and Nginx config from govuk-puppet is a bit of a concern,
particularly for Asset Manager related behaviour which relies quite
heavily on the Nginx config. However, I think this commit is still a
step forward.

[1]: https://github.com/alphagov/govuk-puppet/blob/dd42edb08b978feb95627032211a1084260be2cb/hieradata/common.yaml#L523-L524