Delete all files but not directories in /app/tmp on setup
Previously we were using `rm -rf /app/tmp/*`, but when I introduced the
/app/tmp/uploads & /app/tmp/fake-s3 directories, I didn't want to remove
those directories since they're shared as volumes on the asset-manager
container. However, my changes meant that files in /app/tmp were not
getting cleared up. This in turn meant that an errors.log file from one
build might get left behind and confuse the situation for a subsequent
build. In particular, this code [1] checks for the presence of
errors.log in order to decide whether to report that errors occurred.

I've also excluded `.keep` files from deletion, because otherwise they
show up in `git status` as having been deleted which is confusing.

[1]:
https://github.com/alphagov/publishing-e2e-tests/blob/65af0d50f9bad3814d9b64174f2f63aac89e0d6c/Jenkinsfile#L262-L268