Update to prefer pulling docker images over building
By pulling the deployed-to-production docker tag for the apps
it becomes quicker to get the different applications setup as
we no longer need to wait for bundle installs or OS setup to
occur. This makes the project to be easier to maintain as it
lowers the cost of entry for developers who have not
previously set it up locally or who only occasionally work
on this project.