Save the two step verification secrets to a file
This means the tests can be rerun in the same environment, where the
secret will be provided the first time, and then read from the
filesystem the second time.

This makes it possible to run the tests multiple times within the same
environment, which is useful for developing and testing the tests
themselves.