Group more ENV vars in extension fields
This changes the grouping from being just a singular production one
(that was related to production changes) to a more generic govuk-app one
with a corresponding draft-govuk-app one.

It's a total pain thinking of a common name for a grouping here. I went
with govuk-app rather than common as that felt a little more
appropriate, but I don't think it's a particularly good name.