Fix scenario where attachment doesn't appear in section
We've had a few instances where this error happens:

```
After 60 seconds,

http://draft-origin.dev.gov.uk/guidance/uploading-atttachent-to-manual-a518a3ae
  -9133-4a90-a3cb-037ca7741829/whats-become-of-waring-1516019030

didn't match

Uploading atttachent to manual attachment 0959f5e5-53e7-4194-9e4a-cf67d83591a7

for has_text?
```

By moving this reload_url_until_match to before the attachment gets added we
can ensure that the request to the publishing-api to create the new section
and the request to the publishing-api to add the attachment link happen in
the correct order.

The logs don't provide enough information to confirm this theory, but in my
opinion is the only possible cause of this flakyness.