Group app data in Jenkinsfile
This introduces a list of all the apps which can be a single place for
the many aspects of app data that is used in the Jenkinsfile.

It does a trim when setting commitish as we have had some problems with
rogue new line characters causing problems on the end of a commitish.

I noticed that there wasn't actually deployed-to-production and similar
branches setup for router so that defaults to master.