Calculate which apps we plan to build based on commit
This builds up an env var of apps which aren't on a default commit and
passes them in so they can be run through docker build.

This will mean that for cases where we have a branch that is on docker
hub we will pull the image down and then try build it. I had hoped that
with docker caching this would mean it doing virtually nothing but that
doesn't seem to be the case, so this could use some iteration.