Boot up Whitehall so it can publish and render content
We were missing containers for the frontend and workers of Whitehall.

Creating Whitehall documents also requires rummager to have a line of
communication to publishing-api.

The publishing_api:publish_special_routes rake task configures the
router to send /government requests to whitehall-frontend.

We mount the asset-manager-tmp folder as a volume, because the
whitehall-admin passes assets to the whitehall-worker via it.