Update make setup to reduce false errors reported during setup
Due to the layout of the setup step we found that we were generating a
large amount of errors that weren't valid. By launching the initial
cleanup command with `--no-deps` we eliminate multiple applications
being started before their setup steps have run. We've also added the
`--rm` flag to all of the run commands to reduce the number of one use
containers that are left around after the setup.

The only error that is still generated during setup is
`InvalidFormatError` from the specialist-publisher publishing it's
finders as these appear to no longer be valid.