Add wait-for-it for use with rummager and elasticsearch
As elasticsearch was no longer booted by the cleanup command we had to
add back in starting it in detached mode before running any other
commands. By adding `wait-for-it.sh` in and updating the commands for
rummager and rummager-worker we can instead make the docker wait until
elasticsearch is up on the correct port before running container
commands that depend on it.