Update elasticsearch and elasticsearch config
Rummager utilises multiple features that are not enabled as part of the base
image for elasticsearch. This adds the indexed and inline scripts to the
elasticsearch config ensuring search functions correctly instead of returning
a 503.

For search to be available via the frontend we have to publish the
finders during setup.

We also updated the elasticsearch version used to match the one listed
in puppet.