Set GOVUK_APP_DOMAIN_EXTERNAL in docker-compose.yml
This is new in Plek 2.1.0, and represents the domain used as a
template for external URLS. Applications can use this by using the
`external_url_for` method in Plek.

For the docker-compose environment, this is the same as the
GOVUK_APP_DOMAIN (which is effectively for internal routing), as I'm
assuming that there isn't any different routing by domain for app to
app connections compared to user (or the Publishing E2E tests) to app
connections.