Fix the flaky dialog interaction on Publisher
We were getting a scenario where the test was clicking the
"Skip review" link and progressing to the Publisher homepage.

Instead of clicking the link to open the dialog box we now submit
the form while it is still invisible.

We still need to syncronize each form submission as the form is
put into the DOM at each step even if it is inaccessible.  We use
the link without the disabled class as a way to ensure the document
has progressed to the next state.

Example failure:
https://ci.integration.publishing.service.gov.uk/job/publishing-e2e-tests/job/test-against/4571/