Wait for Preview draft manual page load to finish on attachment test
We've had a flakyness from this test where the find_link(section_title)
was picking up the title from the publishing app page and not the frontend.

By waiting for the frontend page to be loaded we can ensure that it
is picking up the link on the rendered version of the document.

Example failure:
https://ci.integration.publishing.service.gov.uk/job/publishing-e2e-tests/job/test-against/4555/