Click link to open up "New part" section in a loop
There exists a race hazard here between the browser adding the event
handler to the "Add new part" link and Capybara trying to click that
link.

If capybara clicks the link before the event handler is registered
then it won't open up the new part form.

This loop exists to ensure that the "Add new part" link is clicked
:until the form appears.