Wait for databases to be healthy before running commands
To avoid any potential race conditions with the databases starting and
commands being run on them we've added a series of external healthchecks
to ensure they are up and responding to commands before we proceed. This
is required as docker-compose healthchecks no longer allow you to depend
on them being healthy natively.

This rake task has to be run outside of the docker containers as it uses
`docker-compose` commands which are not available on the container.