Ensure all apps are started before we run tests
The end-to-end tests will only work correctly if all the dependent apps
are running. To ensure these are up before running the tests we query the
healthcheck for all services that have one, excluding the test runner
container (publishing-e2e-tests) as that will only be booted when we
start running the tests.

The switch to alpine broke the elasticsearch healthcheck, as it doesn't
have curl installed.  Switch back to a fat elasticsearch image.