Move `expect_rendering_application` above `expect_url_matches_*_gov_uk`
It is possible for PhantomJS to be loading the rendered version, but
for Capybara to have gone past the `expect(page).to have_content`
as the publishing application already has the content.

By moving the `expect_rendering_application` above the call to
`expect_url_matches_live_gov_uk` Capybara will use its built in
repeated find functionality to ensure the frontend has loaded.

https://ci.integration.publishing.service.gov.uk/job/publishing-e2e-tests/job/run-tests-continually/1128/