Use case insensitivity when checking if the page shows "gone"
There exists a race condition where the router will query content-store
and it exists, but when it gets to the government-frontend the document
is gone.

If the router renders the 410 then it displays a message with a
lowercase "gone".

If the government-frontend renders the 410 then it displays it as
"Gone".

By using a case insensitive match we can handle both cases.

Example race condition example:
https://ci.integration.publishing.service.gov.uk/job/publishing-e2e-tests/job/run-tests-continually/1084/