Register a JQuery ready handler for syncronizing Capybara and PhantomJS
For situations where you want to block until the DOM is full loaded the
`wait_for_jquery_ready_event` will allow you to do so.

JQuery won't add the "e2e-tests-loaded" class onto the document until
the DOM is loaded.  We then use Capybaras built in `find` method which
repeatedly looks for that class for the configured timeout period.

For Mainstream publisher adding a part to an artefact this should be
a suitable replacement for the current loop behaviour.  Once the DOM
is loaded the nested_form event on the "Add new part" link will have
been attached.

Going forward this method should be a simple and reusable method to
use in similar situations.