Add clean_apps recipe to Makefile
If you haven't worked on the publishing-e2e-tests in a while your
apps folder can have old versions of the apps.  Removing all the
apps and then cloning again is the best way of ensuring you are
testing against the latest version of the apps.

When running Docker on Linux, such as on Jenkins, it will have
created files inside of the mounted volumes e.g. apps/<app>/logs/
as root. Running the `rm` command outside of docker won't have
the necessary perissions to remove those files.

We're using docker instead of docker-compose in this case as
docker-compose blows up if any of the apps/<app> directories
don't exist.  This allows you to run `clean` without having
first run `clone`.