Add clean_docker command which removes built app images
When you build an app image locally with `docker-compose build <app>`
it will tag the image the imange name inside of the docker-compose.yml

E.g.

   govuk/manuals-publisher:master

In dockers eyes this image is more up to date than the version on docker
hub and any call to docker-compose pull won't update your locally built
image.  To help in this scenario we've added a new make command +
associated rake task to delete all images on your computer with a govuk/
tag and which don't exist in a registry.