Move the timeout up to 20s and interval up to 10s
Reducing the timeout in eb8025f4 had the unforeseen side effect of
increasing the amount of CPU time required to service these
healthcheck requests.  We're now in a balancing act of finding a
frequency of healthchecking which provides a healthy status quickly
vs one which consumes a proportionate amount of system resources.

After much experimentation on both Ryan and Adrians macs, the
combination of 20s timeout + 10s interval was performing well.
At the beginning of the boot up process, the timeouts will be
more common so we believe this is why a longer backoff here gives
enhanced stability.