Wait up to 3 minutes for whitehall admin to start
Whitehall is quite the slow booter and by checking that it's started
soon after starting apps it may not be ready. This pumps the time we
wait up to a rather large 3 minutes.

On my local machine I've seen it vary from about 17 seconds with no
other apps and peaked at around ~1 min. Other people have experienced it
taking past a minute.