Add rudimentary cookie support to the retry helpers
The RetryHelper module methods use a couple of disconnected HTTP
clients, different from the main client used for the bulk of the
testing.

The lack of shared state causes problems when authentication via
Signon is used, as even if the main HTTP client is successful in
logging in, the retry helpers won't be able to access the same
content.

Many of the current tests rely on the retry helpers being able to do
this, the most common example being testing for content on the
draft-origin which would usually require a user to authenticate with
the authenticating-proxy service.

To make this work, manually store and inject the cookies in to the
separate Capybara session, and the HTTParty client used within the
retry helpers module. Poltergeist returns cookies for the domain the
main client is on, so store the cookies in a hash by domain. This is
important, as often the retry helpers are testing a different domain
from which the main client has previously been making requests to.

With this change, I was able to get some of the tests that use the
draft-origin working with the test environment created using
govuk-guix, which uses Signon and Authenticating Proxy.