Mark removing a taxon test as flaky
We are seeing a number of tests fail with messages like this:

Aborting reloading
http://www.dev.gov.uk/removed-taxon-bb2627d6-a21a-4d7e-a802-1b0b32507997
as a 500 was returned

Which seem to have originated since
https://github.com/alphagov/collections/pull/615 was merged - although I'm not
sure how this didn't error in some way before this.

The hypothesis for why this is occurring is that router is updated after
content store and collections is trying to serve an unpublished
document. Once collections can return correct redirect/gone responses
then this test can have the flaky tag removed.