Add Redis to the whitehall-admin config
The new topic tagging flow in Whitehall-admin needs to be able to load the taxonomy from Redis

Also set a REDIS_URL in whitehall-admin and whitehall-worker config so that the taxons are read
from, and documents are published to the same instance of Redis.

The default setting for REDIS is to use a host IP of 127.0.0.1 but we need to connect to the Redis host.