Reload whitehall publication finder once
This test seems quite unreliable, I ran it 30 times without success.
It fails because `Session#find` can't find an <a> with the right text,
which I find strange because I think the `reload_url_until_match`
should make that work.  Whatever the cause, adding in another page
load makes this test pass.

It's not a great solution, but this one test is making developing the
e2e tests really hard, so it'll do for now.