Specify the chromedriver version to download
This file is read by the chromedriver-helper Gem, which downloads
chromedriver when the tests are being run.

This isn't a great way of managing software :( but at least with the
version specified here, we should get this Docker container for
running the tests in back to a workable state, and this might help
prevent similar issues in the future.