Remove Whitehall publications finder from tests
This commit removes the Whitehall publications finder from tests and instead replaces it with the all content finder for those tests to continue to work. The reason behind this is that Whitehall finders are being removed and the publications finder does not directly map to any of the new finders.

Trello: https://trello.com/c/3BuGZK2A