Reset databases rather than setup
Databases were being setup with rake db:setup, which
creates the database, loads the schema and seeds it.
Changing it to rake db:reset means that it drops
the database before creating it, loading the schema
and then seeds data.

The change is desirable because when developing tests
the setup occasionally exits when validations fail and
the lengthy docker startup procedure has to be restarted.
This is a source of pain and frustration and makes
developing tests take longer than necessary.

https://edgeguides.rubyonrails.org/active_record_migrations.html#setup-the-database