Ensure gov.uk/api/organisations exists when seeding contacts-admin
This is to support https://github.com/alphagov/contacts-admin/pull/509
which means that contacts-admin is now using the public organisations
API rather than talking directly to Whitehall.

The root problem is that contacts-admin depends on external services to
seed its own database. Ideally, this wouldn't be the case.

The ordering here is particular. We can't just publish all the routes
before seeding contacts-admin, because contacts-admin itself has some
routes to publish. Instead, we have to publish just the
/api/organisations route, which is rendered by collections, from data in
search-api.