Update rubyLinter call
rubyLinter no longer has a parameter for running lint on diffs. e2e tests
are broken at the linting stage currently because of this.

I'm not sure whether this might cause further problems, but if so, we could
always remove the linting section from here altogether (or do something
else).