Set a password for Postgres
The postgres docker images were changed 3 days with a breaking change
[1]. On starting the container we'd see the following error:

```
Error: Database is uninitialized and superuser password is not specified.
       You must specify POSTGRES_PASSWORD for the superuser. Use
       "-e POSTGRES_PASSWORD=password" to set it in "docker run".

       You may also use POSTGRES_HOST_AUTH_METHOD=trust to allow all connections
       without a password. This is *not* recommended. See PostgreSQL
       documentation about "trust":
       https://www.postgresql.org/docs/current/auth-trust.html
```

To resolve this I've set up a password for postgres and applied that to
all the postgres apps DATABASE_URL environment variables. These apps
were previously relying on the DATABASE_URL constant set in their
individual docker images this, however, overwrites those with one
featuring a password.

[1]: https://github.com/docker-library/postgres/pull/658