Retry if document doesn't appear in search results
We see 'Failure/Error: expect(page).to have_content(title)'
occasionally when checking a new document appears in search
results.

It's possible this is due to search not indexing a document
quickly enough. We don't see this error in other cases
where we retry the search page, as in the Publishing a contact
test.