Add ASSET_HOST env var to frontend apps
This adds environment variables for all the frontend apps that can be
used to specify the hostname used for assets. This is to support the
change of hostname for assets of GOV.UK apps from the
`assets.publishing.service.gov.uk` hostname to the `www.gov.uk` one.
For more information on this change
see: https://github.com/alphagov/govuk-puppet/pull/10360.

To serve assets on an assets hostname in these tests requires a bunch of
nginx configuration to proxy asset requests to the corresponding Rails
application. Rather than moving this configuration to the www hostname I
thought it would be better to take the same approach as govuk-docker and
specify that assets are served by a particular host. This saves the need
for additional nginx configuration.

This change pre-emptively applies this environment variable to all
frontend apps with the expectation that the apps will gradually support
this env var as the apps are migrated. Once all the apps have been
migrated the asset configuration to proxy to apps can then be removed.