Organise test containers depends and links
The depends_on for this container had become somewhat inconsistent
in it's contents (this is mostly a consequence of the containers
needed already being up before this container starts). I've updated it
to depend on things it will talk to while omitting things that are
depended on implicitly (for example router depends on all the frontend
apps).

I've done some organising of the links section too so this is easier to
understand.