Add explicit REDIS_URL env vars to all apps that use Redis
This is to assist with the transition from REDIS_HOST/REDIS_PORT to
REDIS_URL. This is done so all of these apps are set-up with the
necessary env var in here rather than relying on changing the
Dockerfile. It also a fixes a few places where redis dependencies
weren't specified.

We have a somewhat strange pattern where the Dockerfiles for GOV.UK apps
have default env vars that represent services defined here (for example:
`REDIS_HOST: redis`). This was intended to be useful for a few reasons,
but reflecting a couple of years later - it's mostly confusing.