Add new container running MongoDB 2.6.12 for routing
This commit adds a new container running MongoDB `2.6.12`, aimed specifically at Router and Router API. The MongoDB version used for Router / Router API in production is already `2.6.12`, however this test suite previously created a single MongoDB instance running at the higher version `3.4` and used this for all apps requiring MongoDB; this commit fixes that misrepresentation.

The reason it's necessary to make the changes in this commit is because we are now polling the MongoDB replica set inside of Router to determine whether or not to pull new routes. The specifics of how a replica set can be polled differs between MongoDB versions `2.6.12` and `3.4`; the results of each are not interchangeable, which means that we can't simply piggy back on the existing `3.4` instance.