Fix test that relies on router having a presumed state
This test will wait for a recently published piece of content to change
from a 200 response to a 301 redirect. However an assumption of this
test is that the content has made it into router and is returning a 200
response. This assumption, while presumptuous, used to hold until the
recent changes to Router which switched to a polling rather than a
notifying mechanism [1].

To resolve this I've set this wait condition to be able to handle the
state where router hasn't updated yet.

[1]: https://github.com/alphagov/router/pull/195