Download one copy of Chrome per parallel worker
We've started seeing the error

    ChildProcess::LaunchError:
      Text file busy - /root/.webdrivers/chromedriver

Which sounds a lot like
https://github.com/titusfortner/webdrivers/issues/77, but that fix was
applied in 2019 and it solved the issues we had then.  This is a new
occurrence, possibly related to a new version of Chrome (as one came
out recently).

Downloading one copy of chrome for each test runner isn't great, but
it's acceptable if it fixes the issue.