Specify the PORT and DATABASE_URL env var for Whitehall.
This will be needed in order to work with the refactored Dockerfile in
alphagov/whitehall. It also makes sense to specify the port and database url here anyway.

Whitehall-worker to run Sidekiq directly. The reasoning behind this in summary is we removed Foreman as the process runner and replaced it with Puma. for more detail please refer to the below commit;
https://github.com/alphagov/publishing-e2e-tests/commit/1958ec4be24641bbbf3d6d9ff525e40f0509d5ef

https://trello.com/c/ZCL2pDyO/739-package-whitehall