Update Collections Publisher tests to work with design system
We recently updated Collections Publisher to use the GOVUK
Design System. These changes do a few things

1. We no longer use select2 so these fields need to be updated to use
select inputs.
2. There is a subtle behaviour change where clicking on a documents
frontend url to visit the page opens in a new tab. Previously it opened
in the same window
3. Fuzzy searches `published` and `archived`. When selenium translates
the page it sees that the state in the tag has been transformed by the
design system and returns it upcased as `PUBLISHED`. This change
makes it regression proof.